finished d conversion

+++ format_help.c
@@ -6,7 +6,7 @@
 /*   By: jfelty <jfelty@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/12 15:02:08 by jfelty            #+#    #+#             */
-/*   Updated: 2019/10/12 18:50:32 by jfelty           ###   ########.fr       */
+/*   Updated: 2019/10/13 17:45:46 by jfelty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,9 +31,9 @@ int64_t		get_arg_signed(int lmod, va_list args)
 	if (lmod == 0)
 		num = va_arg(args, int32_t);
 	else if (lmod == 1)
-		num = va_arg(args, int8_t);
+		num = (int8_t)va_arg(args, int);
 	else if (lmod == 2)
-		num = va_arg(args, int16_t);
+		num = (int16_t)va_arg(args, int);
 	else if (lmod == 3 || lmod == 4)
 		num = va_arg(args, int64_t);
 	return (num);
@@ -47,9 +47,9 @@ uint64_t		get_arg_unsigned(int lmod, va_list args)
 	if (lmod == 0)
 		num = va_arg(args, uint32_t);
 	else if (lmod == 1)
-		num = va_arg(args, uint8_t);
+		num = (uint8_t)va_arg(args, uint32_t);
 	else if (lmod == 2)
-		num = va_arg(args, uint16_t);
+		num = (uint16_t)va_arg(args, uint32_t);
 	else if (lmod == 3 || lmod == 4)
 		num = va_arg(args, uint64_t);
 	return (num);
